refactor to make convert_to_set reusable
